requirements.txt
feedparser>=6.0
arxiv>=2.1
beautifulsoup4>=4.12
requests>=2.31
openai>=1.40
resend>=2.0
jinja2>=3.1
python-dateutil>=2.8
